Autoformat text using simple-markdown -mG
```zsh
for i in **/*.md; do simple-markdown -mG < "$i" | sponge "$i"; done
```

If this commit appears in your `git blame`, you need to run the
following to configure the `.git-blame-ignore-revs` file.

```
git config blame.ignoreRevsFile .git-blame-ignore-revs
```

If you have done that, then please replace the erroneous commit in that
file with the hash of this commit.